Implemented the route tracking via a Service. It makes it possible to register a route even if the app is in background. Thus, I added GPSService. Now if the user presses Stop when no location was added to the database yet, the app doesn't crash, but it show the user a message to inform him of the issue. No more registration when in pause.
